Inlined the strategy that will no longer be needed
