+++ model/map/src/main/java/org/keycloak/models/map/credential/MapSingleUserCredentialManager.java
@@ -24,24 +24,26 @@ import org.keycloak.credential.CredentialInputValidator;
 import org.keycloak.credential.CredentialModel;
 import org.keycloak.credential.CredentialProvider;
 import org.keycloak.credential.CredentialProviderFactory;
-import org.keycloak.credential.SingleEntityCredentialManagerStrategy;
 import org.keycloak.models.KeycloakSession;
+import org.keycloak.models.ModelDuplicateException;
 import org.keycloak.models.RealmModel;
 import org.keycloak.models.SingleEntityCredentialManager;
 import org.keycloak.models.UserModel;
+import org.keycloak.models.map.user.MapUserCredentialEntity;
 import org.keycloak.models.map.user.MapUserEntity;
 
+import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.stream.Stream;
 
 /**
  * Handling credentials for a given user.
  *
  * This serves as a wrapper to specific strategies. The wrapping code implements the logic for {@link CredentialInputUpdater}s
- * and {@link CredentialInputValidator}s. Storage specific strategies can be added, like for example, in
- * {@link MapSingleEntityCredentialManagerStrategy}.
+ * and {@link CredentialInputValidator}s.
  *
  * @author Alexander Schwartz
  */
@@ -50,7 +52,6 @@ public class MapSingleUserCredentialManager implements SingleEntityCredentialMan
     private final UserModel user;
     private final KeycloakSession session;
     private final RealmModel realm;
-    private final SingleEntityCredentialManagerStrategy strategy;
     private final MapUserEntity entity;
 
     public MapSingleUserCredentialManager(KeycloakSession session, RealmModel realm, UserModel user, MapUserEntity entity) {
@@ -58,7 +59,6 @@ public class MapSingleUserCredentialManager implements SingleEntityCredentialMan
         this.session = session;
         this.realm = realm;
         this.entity = entity;
-        this.strategy = new MapSingleEntityCredentialManagerStrategy(entity);
     }
 
     @Override
@@ -69,7 +69,7 @@ public class MapSingleUserCredentialManager implements SingleEntityCredentialMan
 
         List<CredentialInput> toValidate = new LinkedList<>(inputs);
 
-        strategy.validateCredentials(toValidate);
+        entity.credentialManager().validateCredentials(toValidate);
 
         getCredentialProviders(session, CredentialInputValidator.class)
                 .forEach(validator -> validate(realm, user, toValidate, validator));
@@ -79,7 +79,7 @@ public class MapSingleUserCredentialManager implements SingleEntityCredentialMan
 
     @Override
     public boolean updateCredential(CredentialInput input) {
-        return strategy.updateCredential(input) ||
+        return entity.credentialManager().updateCredential(input) ||
                 getCredentialProviders(session, CredentialInputUpdater.class)
                         .filter(updater -> updater.supportsCredentialType(input.getType()))
                         .anyMatch(updater -> updater.updateCredential(realm, user, input));
@@ -88,45 +88,63 @@ public class MapSingleUserCredentialManager implements SingleEntityCredentialMan
     @Override
     public void updateStoredCredential(CredentialModel cred) {
         throwExceptionIfInvalidUser(user);
-        strategy.updateStoredCredential(cred);
+        entity.getCredential(cred.getId()).ifPresent(c -> {
+            c.setCreatedDate(cred.getCreatedDate());
+            c.setUserLabel(cred.getUserLabel());
+            c.setType(cred.getType());
+            c.setSecretData(cred.getSecretData());
+            c.setCredentialData(cred.getCredentialData());
+        });
     }
 
     @Override
     public CredentialModel createStoredCredential(CredentialModel cred) {
         throwExceptionIfInvalidUser(user);
-        return strategy.createStoredCredential(cred);
+        MapUserCredentialEntity credentialEntity = MapUserCredentialEntity.fromModel(cred);
+
+        if (entity.getCredential(cred.getId()).isPresent()) {
+            throw new ModelDuplicateException("A CredentialModel with given id already exists");
+        }
+
+        entity.addCredential(credentialEntity);
+
+        return MapUserCredentialEntity.toModel(credentialEntity);
     }
 
     @Override
     public boolean removeStoredCredentialById(String id) {
         throwExceptionIfInvalidUser(user);
-        return strategy.removeStoredCredentialById(id);
+        return entity.removeCredential(id);
     }
 
     @Override
     public CredentialModel getStoredCredentialById(String id) {
-        return strategy.getStoredCredentialById(id);
+        return entity.getCredential(id).map(MapUserCredentialEntity::toModel).orElse(null);
     }
 
     @Override
     public Stream<CredentialModel> getStoredCredentialsStream() {
-        return strategy.getStoredCredentialsStream();
+        return Optional.ofNullable(entity.getCredentials()).orElse(Collections.emptyList()).stream()
+                .map(MapUserCredentialEntity::toModel);
     }
 
     @Override
     public Stream<CredentialModel> getStoredCredentialsByTypeStream(String type) {
-        return strategy.getStoredCredentialsByTypeStream(type);
+        return getStoredCredentialsStream()
+                .filter(credential -> Objects.equals(type, credential.getType()));
     }
 
     @Override
     public CredentialModel getStoredCredentialByNameAndType(String name, String type) {
-        return strategy.getStoredCredentialByNameAndType(name, type);
+        return getStoredCredentialsStream()
+                .filter(credential -> Objects.equals(name, credential.getUserLabel()))
+                .findFirst().orElse(null);
     }
 
     @Override
     public boolean moveStoredCredentialTo(String id, String newPreviousCredentialId) {
         throwExceptionIfInvalidUser(user);
-        return strategy.moveStoredCredentialTo(id, newPreviousCredentialId);
+        return entity.moveCredential(id, newPreviousCredentialId);
     }
 
     @Override
